+++ src/capturer/vec.rs
@@ -40,6 +40,14 @@ impl Capturer for VecCapturer {
     &self.monitor
   }
 
+  fn texture(&self) -> &ID3D11Texture2D {
+    &self.texture
+  }
+
+  fn texture_desc(&self) -> &D3D11_TEXTURE2D_DESC {
+    &self.texture_desc
+  }
+
   fn buffer(&self) -> &[u8] {
     &self.buffer
   }